[upd&ref] --divided main to main and window and refactored init and video_load_btn_clicked event in mainwindow
