+++ interface/interface.ui
@@ -358,7 +358,7 @@
      <string>Выберите номер остановочного пункта</string>
     </property>
    </widget>
-   <widget class="QComboBox" name="comboBox">
+   <widget class="QComboBox" name="stops_combo_box">
     <property name="geometry">
      <rect>
       <x>180</x>
